Add AMIC A25LQ64 to internal JEDEC ID database

+++ common/ffsparser.cpp
@@ -4676,10 +4676,9 @@ make_partition_table_consistent:
                         UModelIndex partitionIndex = model->addItem(localOffset + partitions[i].ptEntry.Offset.Offset, Types::CpdPartition, Subtypes::ManifestCpdPartition, name, UString(), info, header, body, UByteArray(), Fixed, parent);
                         
                         // Parse data as extensions area
-			// Add the header size as a local offset
-			// Since the body starts after the
-			// header length
-                        parseCpdExtensionsArea(partitionIndex, header.size());
+                        // Add the header size as a local offset
+                        // Since the body starts after the header length
+                        parseCpdExtensionsArea(partitionIndex, (UINT32)header.size());
                     }
                 }
             }
